requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
Flask-Mail==0.9.1
Werkzeug==3.1.3
python-dotenv==1.0.0
python-socketio==5.10.0
python-engineio==4.8.0
gunicorn==21.2.0
eventlet==0.40.4
